Implemented macdeployqt error handling

+++ scripts/mac-deployqt.sh
@@ -47,6 +47,10 @@ fi
 
 info "Call macdeployqt:"
 ${QT_PATH}/bin/macdeployqt src/bin/Spectrecoin.app -qmldir=src/qt/res -always-overwrite -verbose=2
+rtc=$?
+if [[ ${rtc} != 0 ]] ; then
+    die ${rtc} "Error during macdeployqt!"
+fi
 
 info "Remove openssl 1.0.0 libs:"
 rm -v src/bin/spectrecoin.app/Contents/Frameworks/libssl.1.0.0.dylib
@@ -72,4 +76,8 @@ done
 
 info "Create dmg package:"
 ${QT_PATH}/bin/macdeployqt src/bin/Spectrecoin.app -dmg -always-overwrite -verbose=2
+rtc=$?
+if [[ ${rtc} != 0 ]] ; then
+    die ${rtc} "Error during macdeployqt!"
+fi
 mv src/bin/Spectrecoin.dmg Spectrecoin.dmg
